+++ src/main/java/cart/dao/ItemDao.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.model.Item;
+import cart.domain.item.Item;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
