+++ src/main/java/cart/domain/item/ItemName.java
@@ -1,4 +1,4 @@
-package cart.model;
+package cart.domain.item;
 
 import cart.exception.item.ItemFieldNotValidException;
