+++ src/main/java/cart/domain/item/ItemPrice.java
@@ -1,4 +1,4 @@
-package cart.model;
+package cart.domain.item;
 
 import cart.exception.item.ItemFieldNotValidException;
 import java.math.BigInteger;
