+++ src/test/java/cart/domain/item/ItemNameTest.java
@@ -1,4 +1,4 @@
-package cart.model;
+package cart.domain.item;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
