+++ src/test/java/cart/fixture/ResponseFactory.java
@@ -1,7 +1,7 @@
 package cart.fixture;
 
 import cart.controller.dto.ItemResponse;
-import cart.model.Item;
+import cart.domain.item.Item;
 import cart.service.dto.ItemDto;
 
 public final class ResponseFactory {
